# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Grow5X API
  description: |
    API del backend de Grow5X - Plataforma de trading e inversiones con sistema de referidos y comisiones.

    ## Autenticación
    La mayoría de endpoints requieren autenticación mediante JWT token en el header Authorization:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Roles
    - **user**: Usuario regular con acceso a funciones básicas
    - **admin**: Administrador con acceso a funciones administrativas

    ## Códigos de Error Comunes
    - `VALIDATION_ERROR`: Error de validación de datos
    - `UNAUTHORIZED`: Token inválido o faltante
    - `FORBIDDEN`: Permisos insuficientes
    - `RATE_LIMIT_EXCEEDED`: Límite de requests excedido
    - `INTERNAL_ERROR`: Error interno del servidor
  version: 1.0.0
  contact:
    name: Grow5X Team
    email: support@grow5x.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.grow5x.app
    description: Servidor de producción
  - url: http://localhost:5000
    description: Servidor de desarrollo
paths:
...
  /api/admin/packages:
    get:
      tags:
        - Admin - Packages
      summary: Get packages
      description: Get packages endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
    post:
      tags:
        - Admin - Packages
      summary: Create packages
      description: Create packages endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '201':
          description: Created/Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request payload
      security:
        - bearerAuth: []
...
  /api/admin/withdrawals:
    get:
      tags:
        - Admin - Withdrawals
      summary: Get withdrawals
      description: Get withdrawals endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
...
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Obtener información del usuario actual (alias legacy)
      description: Devuelve la información del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/checkout/start:
    post:
      tags:
        - Checkout
      summary: Create start
      description: Create start endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '201':
          description: Created/Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request payload
      security:
        - bearerAuth: []
  /api/checkout/{orderId}/confirm:
    post:
      tags:
        - Checkout
      summary: Create specific confirm
      description: Create specific confirm endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '201':
          description: Created/Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          description: orderId identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request payload
      security:
        - bearerAuth: []
...
  /api/me:
    get:
      tags:
        - User Profile
      summary: Obtener información del usuario actual (endpoint canónico)
      description: Devuelve la información del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/me/purchases:
    get:
      tags:
        - User Profile
      summary: Obtener mis compras (endpoint canónico)
      description: Lista las compras del usuario autenticado
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Elementos por página
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum:
              - pending
              - hash_submitted
              - confirmed
              - rejected
      responses:
        '200':
          description: Lista de compras
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      purchases:
                        type: array
                        items:
                          $ref: '#/components/schemas/Purchase'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - User Profile
      summary: Create purchases
      description: Create purchases endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '201':
          description: Created/Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request payload
      security:
        - bearerAuth: []
  /api/me/purchases/{id}/confirm:
    post:
      tags:
        - User Profile
      summary: Confirm purchase
      description: Confirm a specific purchase by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Purchase identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Purchase confirmation data
      responses:
        '200':
          description: Purchase confirmed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/packages:
    get:
      tags:
        - Packages
      summary: Obtener lista de paquetes
      description: Devuelve todos los paquetes disponibles
      parameters:
        - name: active
          in: query
          description: Filtrar por paquetes activos
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
            default: 'true'
        - name: visible
          in: query
          description: Filtrar por paquetes visibles
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
            default: 'true'
        - name: sort
          in: query
          description: Ordenar resultados
          schema:
            type: string
            enum:
              - sortOrder
              - price_asc
              - price_desc
              - name
            default: sortOrder
      responses:
        '200':
          description: Lista de paquetes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      packages:
                        type: array
                        items:
                          $ref: '#/components/schemas/Package'
                      count:
                        type: integer
                        example: 3
...
  /api/payments/confirm-hash:
    post:
      tags:
        - Payments
      summary: Confirmar hash de transacción
      description: Envía el hash de transacción para confirmación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - purchaseId
                - transactionHash
              properties:
                purchaseId:
                  type: string
                  example: PUR_20240101_001
                transactionHash:
                  type: string
                  example: '0x1234567890abcdef'
      responses:
        '200':
          description: Hash confirmado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Hash de transacción enviado exitosamente
                  data:
                    type: object
                    properties:
                      purchaseId:
                        type: string
                        example: PUR_20240101_001
                      status:
                        type: string
                        example: hash_submitted
...
  /api/payments/submit:
    post:
      tags:
        - Payments
      summary: Enviar solicitud de pago
      description: Crea una nueva solicitud de pago y asigna wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - packageId
                - quantity
              properties:
                packageId:
                  type: string
                  example: PKG_BASIC_001
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 2
      responses:
        '201':
          description: Solicitud de pago creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Purchase'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/public/packages:
    get:
      tags:
        - Public
      summary: Get packages
      description: Get packages endpoint
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Response data
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del endpoint de login.
        Formato: `Bearer <token>`
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          example: USR_20240101_001
        email:
          type: string
          format: email
          example: usuario@ejemplo.com
        firstName:
          type: string
          example: Juan
        lastName:
          type: string
          example: Pérez
        phone:
          type: string
          example: '+1234567890'
        country:
          type: string
          example: México
        role:
          type: string
          enum:
            - user
            - admin
          example: user
        referralCode:
          type: string
          example: REF123456
        isActive:
          type: boolean
          example: true
        isVerified:
          type: boolean
          example: true
        balances:
          type: object
          properties:
            available:
              type: number
              example: 1500.5
            pending:
              type: number
              example: 200
            total:
              type: number
              example: 1700.5
        referralStats:
          type: object
          properties:
            directReferrals:
              type: integer
              example: 5
            totalReferrals:
              type: integer
              example: 15
            totalCommissions:
              type: number
              example: 750
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Package:
      type: object
      properties:
        packageId:
          type: string
          example: PKG_BASIC_001
        name:
          type: string
          example: Paquete Básico
        price:
          type: number
          example: 1000
        currency:
          type: string
          example: USDT
        dailyBenefitRate:
          type: number
          example: 0.015
        dailyBenefitPercentage:
          type: string
          example: '1.5'
        benefitDays:
          type: integer
          example: 30
        totalCycles:
          type: integer
          example: 12
        totalDurationDays:
          type: integer
          example: 360
        totalBenefitAmount:
          type: number
          example: 540
        totalROI:
          type: number
          example: 1.54
        commissionRates:
          type: object
          properties:
            directReferralRate:
              type: number
              example: 0.1
              description: Tasa de comisión por referido directo
            parentBonusRate:
              type: number
              example: 0.1
              description: Tasa de bono por activación de referido
        description:
          type: string
          example: Paquete ideal para principiantes
        features:
          type: array
          items:
            type: string
          example:
            - Beneficios diarios
            - Comisiones por referidos
            - Soporte 24/7
        minPurchase:
          type: integer
          example: 1
        maxPurchase:
          type: integer
          example: 10
        isActive:
          type: boolean
          example: true
        isVisible:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    Purchase:
      type: object
      properties:
        purchaseId:
          type: string
          example: PUR_20240101_001
        user:
          type: object
          properties:
            userId:
              type: string
              example: USR_20240101_001
            email:
              type: string
              example: usuario@ejemplo.com
        package:
          type: object
          properties:
            packageId:
              type: string
              example: PKG_BASIC_001
            name:
              type: string
              example: Paquete Básico
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          example: 1000
        totalAmount:
          type: number
          example: 2000
        currency:
          type: string
          example: USDT
        status:
          type: string
          enum:
            - pending
            - hash_submitted
            - confirmed
            - rejected
          example: pending
        assignedWallet:
          type: object
          properties:
            address:
              type: string
              example: TQn9Y2khEsLJW1ChVWFMSMeRDow5KcbLSE
            network:
              type: string
              example: TRC20
        transactionHash:
          type: string
          nullable: true
          example: '0x1234567890abcdef'
        paymentDeadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error de validación
        code:
          type: string
          example: VALIDATION_ERROR
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email inválido
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
        totalCount:
          type: integer
          example: 100
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...